requirements.txt
absl-py==1.2.0
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.21.4
blinker==1.4
blis==0.7.8
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
colored==1.4.3
config==0.5.1
cssselect==1.1.0
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
dateparser==1.1.1
decorator==5.1.1
demjson3==3.0.5
distlib==0.3.4
emoji==1.6.3
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.0/en_core_web_md-3.4.0-py3-none-any.whl
executing==0.8.3
facebook-scraper==0.2.57
fake-useragent==0.1.11
filelock==3.7.1
Flask==2.1.3
Flask-API==2.0
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
flatbuffers==1.12
fonttools==4.34.4
funcy==1.17
future==0.18.2
gast==0.4.0
gensim==4.2.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gunicorn==20.1.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
libclang==14.0.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
MouseInfo==0.1.3
murmurhash==1.0.7
nltk==3.7
numexpr==2.8.3
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parse==1.19.0
parsel==1.6.0
parso==0.8.3
pathy==0.6.2
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
preshed==3.0.6
prompt-toolkit==3.0.30
protobuf==3.19.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pydantic==1.9.1
pyee==8.2.2
PyGetWindow==0.0.9
Pygments==2.12.0
pyLDAvis==3.3.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pyppeteer==1.0.2
pyquery==1.4.3
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.1
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
selectorlib==0.16.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.0
spacy-legacy==3.0.9
spacy-loggers==1.0.3
spotipy==2.19.0
srsly==2.4.3
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
text2emotion==0.0.5
thinc==8.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
virtualenv==20.14.1
w3lib==1.22.0
wasabi==0.9.1
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
